Remove intermediate error types in module children.
This replaces the intermediate errors SpawnError and WaitError with
simple temporary wrapper types SpawnErrorTag and WaitErrorTag, and
completely removes the ChildFailed type. We are now back to something
much more similar to the original solution, but also much more
ergonomic. The next step is making the various `children` types generic
over their `name` and `argument` again.