Tweak error output.
This indents the "-> reason" explanations of chained errors by two
spaces for better separation. It also removes the "error" addition from
"not all scenarios executed successfully", because that's technically
not an error -- only a consequence of the errors happening before.

This also corrects the integration tests associated with these changes.