Make Display impl of NotUtf8 independent of Debug.
Until now, the Display impl of trytostr::NotUtf8 relied on the Debug
impl of OsString. Because Debug is only meant for programmer-facing
output, and not user-facing output, this is probably a bad idea. Hence,
we instead use the Display impl of String and convert OsStr to String
via OsStr::to_string_lossy().